Completed all endpoints and started debugging on postman, Need to figure out how to implement a delay between all findDuck post requests for each user.
Migrated to the h2 database.

+++ Backend/src/main/java/com/revature/Project1/controllers/MainpageController.java
@@ -11,6 +11,7 @@ import com.revature.Project1.services.WorldService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.HashMap;
@@ -35,34 +36,20 @@ public class MainpageController {
         this.supFunctions = supFunctions;
     }
 
-    @GetMapping(value = "/{id}")
-    public ResponseEntity getUserInfoById(@PathVariable Integer id){
-        Optional<User> resultUser = userService.getUserById(id);
-        if(resultUser.isEmpty()) return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Invalid Id");
+    @GetMapping(value = "")
+    public ResponseEntity getUserInfoById(@CookieValue(value = "project1LoginCookie", defaultValue = "none") String cookie){
+        if(cookie.equals("none")) return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Unauthorized");
+        Optional<User> resultUser = userService.getUserById(Integer.getInteger(cookie));
+        if(resultUser.isEmpty()) return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Unauthorized");
         return ResponseEntity.status(HttpStatus.OK).body(resultUser);
     }
 
-    @GetMapping(value = "/{id}/ducks")
-    public ResponseEntity getDucksOwnedById (@PathVariable Integer id){
-        List<Duck> resultDucks = duckService.getDucksByForeignId(id);
-        return ResponseEntity.status(HttpStatus.OK).body(resultDucks);
-    }
-
-    @GetMapping(value = "/world")
-    public ResponseEntity getWorldInfo(){
-        try {
-            World world = worldService.getWorldValuesById(1);
-            return ResponseEntity.status(HttpStatus.OK).body(world);
-        }
-        catch (Exception e){
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Invalid Id");
-        }
-    }
-
-    @PostMapping(value = "/{id}")
-    public ResponseEntity postNewDuck(@PathVariable Integer id){
-        Optional<User> resultUser = userService.getUserById(id);
-        if(resultUser.isEmpty()) return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Invalid Id");
+//    @Scheduled(fixedRate = 10000)
+    @PostMapping(value = "")
+    public ResponseEntity postNewDuck(@CookieValue(value = "project1LoginCookie", defaultValue = "none") String cookie){
+        if(cookie.equals("none")) return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Unauthorized");
+        Optional<User> resultUser = userService.getUserById(Integer.getInteger(cookie));
+        if(resultUser.isEmpty()) return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Unauthorized");
 
         World world = null;
         try {
@@ -78,21 +65,36 @@ public class MainpageController {
         duckRank = duckAvailable ? duckRank : "C";
 
         Duck returnDuck = new Duck(duckRank);
-        returnDuck.setReference_id(id);
+        returnDuck.setReference_id(Integer.getInteger(cookie));
         returnDuck.setNickname("DefaultName");
 
         try{
+            World updatedWorld = worldService.getWorldValuesById(1);
+            switch(duckRank){
+                case "B": updatedWorld.setB_rank(updatedWorld.getB_rank() - 1);
+                case "A": updatedWorld.setA_rank(updatedWorld.getA_rank() - 1);
+                case "S": updatedWorld.setS_rank(updatedWorld.getS_rank() - 1);
+                case "SS": updatedWorld.setSs_rank(updatedWorld.getSs_rank() - 1);
+            }
+            worldService.setWorldValues(updatedWorld);
+        } catch (Exception e) {
+            return ResponseEntity.status(500).body("Something Went Wrong");
+        }
+
+        try{
             returnDuck = duckService.createDuck(returnDuck);
+
             return ResponseEntity.status(HttpStatus.OK).body(returnDuck);
         }
         catch (Exception e){
             return ResponseEntity.status(500).body("Something Went Wrong");
         }
-
     }
 
-    @PatchMapping(value = "/{id}")
-    public ResponseEntity patchDuckNicknameById (@PathVariable Integer id, @RequestBody Duck duck) {
+    @PatchMapping(value = "")
+    public ResponseEntity patchDuckNicknameById (@CookieValue(value = "project1LoginCookie", defaultValue = "none") String cookie, @RequestBody Duck duck) {
+        if(cookie.equals("none")) return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Unauthorized");
+        if(!cookie.equals(Integer.toString(duck.getReference_id()))) return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Unauthorized");
         try {
             Optional<Duck> resultDucks = duckService.setDuckNicknameById(duck);
             return ResponseEntity.status(HttpStatus.OK).body(resultDucks);
@@ -103,14 +105,22 @@ public class MainpageController {
         }
     }
 
-    @PatchMapping(value = "/{id}/ducks")
-    public ResponseEntity patchSetAvailableDucks(@PathVariable Integer id,@RequestBody World world){
+    @GetMapping(value = "ducks")
+    public ResponseEntity getDucksOwnedById (@CookieValue(value = "project1LoginCookie", defaultValue = "none") String cookie){
+        if(cookie.equals("none")) return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Unauthorized");
+        List<Duck> resultDucks = duckService.getDucksByForeignId(Integer.getInteger(cookie));
+        return ResponseEntity.status(HttpStatus.OK).body(resultDucks);
+    }
+
+    @PatchMapping(value = "ducks")
+    public ResponseEntity patchSetAvailableDucks(@CookieValue(value = "project1LoginCookie", defaultValue = "none") String cookie,@RequestBody World world){
+        if(cookie.equals("none")) return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Unauthorized");
         try {
-           Optional<User> resultUser = userService.getUserById(id);
-           if(resultUser.isEmpty())  return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Improper Info");
-           if(resultUser.get().getAdmin() == 0) return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Unauthorized");
-           World resultWorld = worldService.setWorldValues(world);
-           return ResponseEntity.status(HttpStatus.OK).body(resultWorld);
+            Optional<User> resultUser = userService.getUserById(Integer.getInteger(cookie));
+            if(resultUser.isEmpty())  return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Improper Info");
+            if(resultUser.get().getAdmin() == 0) return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Unauthorized");
+            World resultWorld = worldService.setWorldValues(world);
+            return ResponseEntity.status(HttpStatus.OK).body(resultWorld);
         } catch (ClientSideException e) {
             return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Improper Info");
         } catch (Exception e) {
@@ -118,11 +128,13 @@ public class MainpageController {
         }
     }
 
-    @DeleteMapping(value = "/{id}/ducks")
-    public ResponseEntity deleteDuckById (@PathVariable Integer id, @RequestBody Duck duck){
+    @DeleteMapping(value = "")
+    public ResponseEntity deleteDuckById (@CookieValue(value = "project1LoginCookie", defaultValue = "none") String cookie, @RequestBody Duck duck){
+        if(cookie.equals("none")) return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Unauthorized");
         try {
             Optional<Duck> resultDuck = duckService.deleteDuckById(duck);
-            if(resultDuck.get().getReference_id() != id) return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Improper Info");
+            if (resultDuck.isEmpty()) return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Improper Info");
+            if(resultDuck.get().getReference_id() != Integer.getInteger(cookie)) return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Improper Info");
             return ResponseEntity.status(HttpStatus.OK).body(resultDuck);
         }
         catch (ClientSideException e){
@@ -131,4 +143,46 @@ public class MainpageController {
             return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Something Went wrong");
         }
     }
+
+    @GetMapping(value = "world")
+    public ResponseEntity getWorldInfo(){
+        try {
+            World world = worldService.getWorldValuesById(1);
+            return ResponseEntity.status(HttpStatus.OK).body(world);
+        }
+        catch (Exception e){
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Invalid Id");
+        }
+    }
+
+    @GetMapping(value = "price")
+    public ResponseEntity getBackpackPrice(@CookieValue(value = "project1LoginCookie", defaultValue = "none") String cookie){
+        if(cookie.equals("none")) return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Unauthorized");
+        try {
+            Optional<User> resultUser = userService.getUserById(Integer.getInteger(cookie));
+            if (resultUser.isEmpty()) return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Improper Info");
+            double backpackPrice = Math.pow(1.5,resultUser.get().getBackpack_space() - 5.0);
+            backpackPrice = Math.round(backpackPrice * 100.0) / 100.0;
+            return ResponseEntity.status(HttpStatus.OK).body(backpackPrice);
+        } catch (Exception e){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Unauthorized");
+        }
+    }
+
+    @PatchMapping(value = "price")
+    public ResponseEntity buyBackpack(@CookieValue(value = "project1LoginCookie", defaultValue = "none") String cookie){
+        if(cookie.equals("none")) return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Unauthorized");
+        try {
+            Optional<User> resultUser = userService.getUserById(Integer.getInteger(cookie));
+            if (resultUser.isEmpty()) return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Improper Info");
+            double backpackPrice = Math.pow(1.5,resultUser.get().getBackpack_space() - 5.0);
+            backpackPrice = Math.round(backpackPrice * -100.0) / 100.0;
+            User user = userService.setUserBankAccount(resultUser.get(),backpackPrice);
+            return ResponseEntity.status(HttpStatus.OK).body(user);
+        } catch (Exception e) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Unauthorized");
+        }
+    }
+
+
 }
